Check if bigWig/bigBed file handle is still open, before reading/writing.
Check if bigWig/bigBed file handle is still open, before trying to
read or write data from the bigWig or bigBed file.

This prevents segfaults when the bigWig object is used after it was closed.

    >>> import pyBigWig
    >>> bb = pyBigWig.open('test.bw')
    >>> bb.header()
    {'maxVal': 2L,
     'minVal': 1L,
     'nBasesCovered': 15560789L,
     'nLevels': 6L,
     'sumData': 15611000L,
     'sumSquared': 15711422L,
     'version': 4L}
    >>> bb.close()
    >>> bb.header()
    Segmentation fault (core dumped)

After this patch, the call to functions that require an open bigWig object,
return something like this:

    >>> bb.header()
    RuntimeError: bigWig file handle is not open.